+++ vck-rqes/src/commonMain/kotlin/at/asitplus/wallet/lib/rqes/helper/SimpleQtspAuthorizationService.kt
@@ -12,14 +12,27 @@ import at.asitplus.wallet.lib.oidvci.OAuth2DataProvider
  * Simple wrapper for [SimpleAuthorizationService] which uses [QtspAuthorizationServiceStrategy]
  */
 class SimpleQtspAuthorizationService private constructor(
-    private val authorizationService: SimpleAuthorizationService
+    private val authorizationService: SimpleAuthorizationService,
 ) : OAuth2AuthorizationServerAdapter by authorizationService, AuthorizationService by authorizationService {
+    @Deprecated("Use constructor without `dataProvider` instead")
     constructor(
         dataProvider: OAuth2DataProvider,
         acceptedCredentials: Collection<ConstantIndex.CredentialScheme>,
     ) : this(
         authorizationService = SimpleAuthorizationService(
             dataProvider = dataProvider,
+            strategy = QtspAuthorizationServiceStrategy(
+                authorizationServiceStrategy = CredentialAuthorizationServiceStrategy(
+                    acceptedCredentials.toSet()
+                )
+            )
+        )
+    )
+
+    constructor(
+        acceptedCredentials: Collection<ConstantIndex.CredentialScheme>,
+    ) : this(
+        authorizationService = SimpleAuthorizationService(
             strategy = QtspAuthorizationServiceStrategy(
                 authorizationServiceStrategy = CredentialAuthorizationServiceStrategy(
                     acceptedCredentials.toSet()